requirements.txt
asgiref==3.3.1
certifi==2020.12.5
chardet==4.0.0
Django==3.1.7
idna==2.10
Jinja2==2.11.3
MarkupSafe==1.1.1
pip==21.0.1
psycopg2==2.8.6
pytz==2021.1
requests==2.25.1
setuptools==54.2.0
sqlparse==0.4.1
urllib3==1.26.4
dj-database-url==0.5.0
whitenoise==5.2.0
gunicorn==20.1.0
djangorestframework==3.12.4
django_environ==0.4.5
python-dotenv==0.16.0
docker==4.4.4
